adds show command; refactor commands
