+++ README.md
@@ -1,4 +1,9 @@
 # bm
 ## A command line bookmarking tool
 
-`bm insert --category={category} {link-title} {url}`
+`bm insert [--category={category}] [--interactive={bool}] {link-title} {url}`
+`bm update [--category={category}] {link-title}`
+`bm list [--category={category}]`
+`bm show [--category={category}] {link-title}`
+`bm remove --category={category} {link-title}`
+
