+++ pkg/cmd/remove.go
@@ -1,4 +1,4 @@
-package remove
+package cmd
 
 import (
 	"fmt"
@@ -8,24 +8,10 @@ import (
 	"github.com/spf13/cobra"
 
 	"github.com/evcraddock/bm/pkg/bookmarks"
-	"github.com/evcraddock/bm/pkg/config"
 )
 
-type RemoveOptions struct {
-	Out      io.Writer
-	Config   *config.Config
-	Category string
-	Title    string
-}
-
-func NewRemoveOptions(out io.Writer) *RemoveOptions {
-	return &RemoveOptions{
-		Out: out,
-	}
-}
-
 func NewCmdRemove(out io.Writer) *cobra.Command {
-	o := NewRemoveOptions(out)
+	o := NewBaseOptions(out)
 	cmd := &cobra.Command{
 		Use: "remove",
 		Run: func(cmd *cobra.Command, args []string) {
@@ -39,22 +25,7 @@ func NewCmdRemove(out io.Writer) *cobra.Command {
 	return cmd
 }
 
-func (o *RemoveOptions) prepare(cmd *cobra.Command, args []string) {
-	if category, err := cmd.Flags().GetString("category"); err == nil {
-		o.Category = category
-	}
-
-	cfg, err := config.LoadConfigFile()
-	if err != nil {
-		fmt.Fprintf(o.Out, "%s", err.Error())
-		os.Exit(1)
-	}
-
-	o.Config = cfg
-	o.Title = args[0]
-}
-
-func (o *RemoveOptions) removeBookmark() {
+func (o *baseOptions) removeBookmark() {
 	manager := bookmarks.NewBookmarkManager(o.Config, false, o.Category)
 	err := manager.Remove(o.Title)
 	if err != nil {
