+++ pkg/cmd/show.go
@@ -1,0 +1,43 @@
+package cmd
+
+import (
+	"fmt"
+	"io"
+	"os"
+
+	"github.com/spf13/cobra"
+	"gopkg.in/yaml.v2"
+
+	"github.com/evcraddock/bm/pkg/bookmarks"
+)
+
+func NewCmdShow(out io.Writer) *cobra.Command {
+	o := NewBaseOptions(out)
+	cmd := &cobra.Command{
+		Use: "show",
+		Run: func(cmd *cobra.Command, args []string) {
+			o.prepare(cmd, args)
+			o.showBookmark()
+		},
+	}
+
+	cmd.Flags().String("category", "readlater", "category folder")
+
+	return cmd
+}
+
+func (o *baseOptions) showBookmark() {
+	manager := bookmarks.NewBookmarkManager(o.Config, false, o.Category)
+	bookmarkLocation := manager.GetBookmarkLocation(o.Title)
+	bookmark, err := manager.Load(bookmarkLocation)
+	if err != nil {
+		fmt.Fprintf(o.Out, "%s", err.Error())
+		os.Exit(1)
+	}
+
+	b, _ := yaml.Marshal(bookmark)
+	if _, err := o.Out.Write(b); err != nil {
+		fmt.Fprintf(o.Out, "%s", err.Error())
+		os.Exit(1)
+	}
+}
